+++ CHANGELOG.md
@@ -48,6 +48,7 @@ and this project adheres to
 - 🚑️(agent) allow Mistral w/ vLLM & tools #36
 - ✨(web-search) add Brave search tool #47
 - ✨(models) add mistral support & customization #51
+- 🐛(web-search) add summarization to Brave results #58
 
 
 [unreleased]: https://github.com/numerique-gouv/conversations/compare/HEAD...main